put mains and roots together
